Improve `cloudfront` service in order to update the cloudfront distribution on AWS when any config is changed.
